Traffic_Model1/emergency-routing-model1/requirements.txt
# Core ML
torch>=2.0.0
# torch-geometric>=2.3.0  # Optional — fallback GCN is included
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0

# Data
pandas>=2.0.0
pyarrow>=12.0.0
# osmnx>=1.6.0  # Optional — synthetic graph fallback available
networkx>=3.1
scipy>=1.11.0

# APIs and HTTP
requests>=2.31.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Training utilities
pyyaml>=6.0
tabulate>=0.9.0

# Testing
pytest>=7.4.0
httpx>=0.24.0

# Visualisation
matplotlib>=3.7.0
seaborn>=0.12.0
